requirements.txt
numpy==1.24.3
opencv_contrib_python==4.11.0.86
opencv_python_headless==4.10.0.84
Pillow==9.4.0
Pillow==11.1.0
streamlit==1.40.2
